refactor: 클래스 패키지 이동
- 프로젝트 정의서에 따라 클래스의 패키지 이동
